errors: Create separated module for the exceptions
